schemas/pyproject.toml
[tool.poetry]
name = "mozilla-nimbus-schemas"
version = "2024.10.1"
description = "Schemas used by Mozilla Nimbus and related projects."
authors = ["mikewilli"]
license = "MPL 2.0"
readme = "README.md"
packages = [{ include = "mozilla_nimbus_schemas" }]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2"
polyfactory = "^2.7.2"
typing-extensions = ">=4.0.1" # Required until Python 3.11

[tool.poetry.group.dev.dependencies]
ruff = ">=0.5.0,<0.6.2"
black = ">=23.3,<25.0"
pytest = "^7.3.1"
twine = "^5.1.1"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.ruff]
# # Enable Pyflakes `E` and `F` codes by default.
lint.select = ["F", "E", "W", "I", "N", "YTT", "A", "C4", "RET", "SIM"]
lint.ignore = [
    "A003",
    "E402",
    "E741",
    "F403",
    "N802",
    "N803",
    "N806",
    "N812",
    "N815",
    "RET503",
    "RET504",
    "RET505",
    "SIM102",
]
line-length = 90
[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]

[tool.flake8]
max-line-length = 90
[tool.black]
line-length = 90
